+++ packages/browser-wallet/src/popup/pages/Account/TransactionLog/i18n/da.ts
@@ -1,0 +1,13 @@
+import type en from './en';
+
+const t: typeof en = {
+    header: 'Transaktionsdetaljer',
+    fromAddress: 'Afsender adresse',
+    toAddress: 'Modtager adresse',
+    transactionHash: 'Transaktionens hash',
+    blockHash: 'Blok hash ',
+    events: 'Hændelser',
+    rejectReason: 'Afvisningsårsag',
+};
+
+export default t;