+++ packages/browser-wallet/src/popup/shell/i18n/locales/da.ts
@@ -11,7 +11,8 @@ import visualSettings from '@popup/pages/VisualSettings/i18n/da';
 import about from '@popup/pages/About/i18n/da';
 import addAccount from '@popup/pages/AddAccount/i18n/da';
 import identityIssuance from '@popup/pages/IdentityIssuance/i18n/da';
-import login from '@popup/pages/Login/i18n/en';
+import login from '@popup/pages/Login/i18n/da';
+import transactionLog from '@popup/pages/Account/TransactionLog/i18n/da';
 import changePasscode from '@popup/pages/ChangePasscode/i18n/en';
 
 import type en from './en';
@@ -31,6 +32,7 @@ const t: typeof en = {
     identityIssuance,
     about,
     login,
+    transactionLog,
     changePasscode,
 };