+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   Login screen.
 -   Change passcode settings page.
+-   Transaction details view when clicking a transaction in the transaction list.
 
 ### Changed